1.4 (08 March 2015)
- I understand the skin is well-integrated, but also inseparable where
it is very difficult to customize. I'm in the process of modularizing
key parts so that they are more accessible to both users and skin
authors. It will take some time, but I hope your patience will be
fulfilled for all the new exciting things to come in the future. Cheers!
- Renamed "Fountain of Colors" to "Ensemble"
- Moved album art display function outside of ColorFactory.lua
- Fixed album art display initial width padding